+++ src/cpu/simple_reorder.hpp
@@ -1247,7 +1247,7 @@ struct simple_reorder_impl<SIMPLE_REORDER_TEMPL_CALL,
             }
 
             if (rem_elems != 0 && ithr == nthr - 1){
-                for (int e = nelems - rem_elems; e < nelems; ++e){
+                for (size_t e = nelems - rem_elems; e < nelems; ++e){
                     output[e] = data_t<type_o>((alpha == 1.0 && beta == 0.0) ?
                             input[e]
                             : alpha * input[e] + (beta ? beta * output[e] : 0));
@@ -1381,14 +1381,14 @@ struct simple_reorder_impl<SIMPLE_REORDER_TEMPL_CALL,
         if (alpha == 1.0 && beta == 0.0) {
 #           pragma omp parallel for schedule(static)
 #           pragma simd
-            for (int e = 0; e < nelems; ++e) {
+            for (size_t e = 0; e < nelems; ++e) {
                 output[output_d.off_l(e)] =
                     data_t<type_o>(input[input_d.off_l(e)]);
             }
         } else {
 #           pragma omp parallel for schedule(static)
 #           pragma simd
-            for (int e = 0; e < nelems; ++e) {
+            for (size_t e = 0; e < nelems; ++e) {
                 output[output_d.off_l(e)] = data_t<type_o>(
                     alpha * input[input_d.off_l(e)]
                     + (beta ? beta * output[output_d.off_l(e)] : 0));
